Modifications of the json files, of the use cases and other minor modifications (#59)
* Added the altitude of a point in the json file.

* Minor modifications of some parameters.

* Added a new json which contains data representing a Sentinel-2 satellite.

* First version of the use case 1.

* Minor modifications of the usecase1.

* Corrected mistake : conversion from degree to radian.

* Modified some json files, added a new json file and added a readme file for the json files.

* Added two new use cases for tests (a sentinel-2 and a galileo like constellations).

* Minor modifications.

+++ project/src/main/java/decisionVector/DecisionVector1.java
@@ -95,9 +95,8 @@ public class DecisionVector1 extends DecisionVector {
 		// the anomaly is the only parameter which can change in the use case 1,
 		// so we have one value for every satellite
 		// read the anomaly and add each satellite to the constellation
-		for (Double i = 0.0; i < nbSat; i++) {
-			String nameAnomaly = "anomaly" + Double.toString(i);
-
+		for (Integer i = 0; i < nbSat; i++) {
+			String nameAnomaly = "anomaly" + Integer.toString(i);
 			Double anomaly = (Double) values.get(getIndex(nameAnomaly));
 
 			constellation.addSatellite(a, eccentricity, inclination, rightAscendingNode, periapsisArgument, anomaly,
